Move url validation to utils
